created seperate array api module
