+++ sklearnex/_device_offload.py
@@ -19,16 +19,16 @@ from functools import wraps
 from onedal._device_offload import (
     _copy_to_usm,
     _extract_array_attr,
-    _from_dlpack,
     _get_global_queue,
-    _is_numpy_namespace,
     _transfer_to_host,
     dpnp_available,
 )
 
+from onedal.utils._array_api import _from_dlpack, _is_numpy_namespace
+
 if dpnp_available:
     import dpnp
-    from onedal._device_offload import _convert_to_dpnp, _from_dlpack
+    from onedal.utils._array_api import _convert_to_dpnp
 
 
 from ._config import get_config
